refactor(github): update workflow run approval method to use int for runID
- Changed the type of `runID` in the `ApproveWorkflowRun` method and its interface from `int64` to `int` for consistency.
- Updated the corresponding test case to reflect this change.
- Added a new method in the fake GitHub client to support the updated approval functionality.

These modifications streamline the handling of workflow run approvals within the GitHub client.

+++ pkg/github/client.go
@@ -284,7 +284,7 @@ type Client interface {
 	Used() bool
 	TriggerGitHubWorkflow(org, repo string, id int) error
 	TriggerFailedGitHubWorkflow(org, repo string, id int) error
-	ApproveWorkflowRun(org, repo string, runID int64) error
+	ApproveWorkflowRun(org, repo string, runID int) error
 	GetPendingApproveActionRunsByHeadSHA(org, repo, headSHA string) ([]WorkflowRun, error)
 }
 
@@ -2127,7 +2127,7 @@ func (c *client) TriggerFailedGitHubWorkflow(org, repo string, id int) error {
 // ApproveWorkflowRun approves a workflow run
 //
 // See https://docs.github.com/en/rest/actions/workflow-runs?apiVersion=2022-11-28#approve-a-workflow-run-for-a-fork-pull-request
-func (c *client) ApproveWorkflowRun(org, repo string, runID int64) error {
+func (c *client) ApproveWorkflowRun(org, repo string, runID int) error {
 	durationLogger := c.log("ApproveWorkflowRun", org, repo, runID)
 	defer durationLogger()
